Moved validation classes in their own files
