+++ src/BullOak.Repositories/Session/AlwaysPassValidator.cs
@@ -1,0 +1,9 @@
+﻿namespace BullOak.Repositories.Session
+{
+    internal class AlwaysPassValidator<T> : IValidateState<T>
+    {
+        /// <inheritdoc />
+        public ValidationResults Validate(T state)
+            => ValidationResults.Success();
+    }
+}
