+++ src/BullOak.Repositories/Session/BusinessException.cs
@@ -1,0 +1,11 @@
+﻿namespace BullOak.Repositories.Session
+{
+    using System;
+
+    public class BusinessException : Exception
+    {
+        public BusinessException(string message)
+            :base(message)
+        { }
+    }
+}
