+++ src/BullOak.Repositories/Session/IValidateState.cs
@@ -1,0 +1,7 @@
+﻿namespace BullOak.Repositories.Session
+{
+    public interface IValidateState<in TState>
+    {
+        ValidationResults Validate(TState state);
+    }
+}
