+++ src/BullOak.Repositories/Session/IValidationError.cs
@@ -1,0 +1,10 @@
+﻿namespace BullOak.Repositories.Session
+{
+    using System;
+
+    public interface IValidationError
+    {
+        string Message { get; }
+        Exception GetAsException();
+    }
+}
